Correct type matching

+++ weight-module/src/main/java/app/web/weightModule/application/dto/ReportSummaryDto.java
@@ -13,7 +13,7 @@ class ReportSummaryDto implements ReportSummary {
     private final float totalMaterialWeight;
     private final float weightDifference;
     private final float correctProductPercent;
-    private final float incorrectProductPcs;
+    private final long incorrectProductPcs;
     private final long overFilledProductPcs;
     private final long notFulFilledProductPcs;
 }
